test/integration/apiserver/apply: Move benchmarks to their own file
